chore(bal): split index into partial files
The lane infrastructure (LaneStore, LaneBase, Lane<TValue>, StorageLane,
LaneSpans, MutableBookkeeping, StorageScratch, PooledArrays, plus the
FillRowStarts and PartitionInQuarters helpers) moves into
BlockAccessListValidationIndex.LaneStore.cs as a partial-class file.

* The main file shrinks from 1231 to ~520 lines and stays focused on
  the index's public API: Build, Add, ChangesEqual, FindStructuralMismatch,
  HasStorageReadsForOrdinal, the AddressIndex helper, etc.
* The new file (~520 lines) is a coherent unit on its own — every nested
  type stays at the same visibility (private nested for everything except
  LaneStore and LaneSpans which were already internal so the validation
  slow path can reach them).

Pure file-organisation change. No behavioural changes; tests unchanged.
